Adding registry for AWS Cloudwatch
Cloudwatch classes already existed in micrometer so now we are just
exposing them and setting up properties for micronaut to be able to
utilize them.

Difference between this and other implementations is this has a required
field of `namespace`. The default has been set to `micronaut` and it can
be customized via config.